+++ packages/map-template/src/components/MapTemplate/MapTemplate.jsx
@@ -6,7 +6,7 @@ import i18n from 'i18next';
 import initI18n from '../../i18n/initialize.js';
 import './MapTemplate.scss';
 import { mapClickActions } from '../../constants/mapClickActions.js';
-import MapWrapper from "../MapWrapper/MapWrapper";
+import MapWrapper from '../MapWrapper/MapWrapper';
 import SplashScreen from '../SplashScreen/SplashScreen';
 import VenueSelector from '../VenueSelector/VenueSelector';
 import BottomSheet from '../BottomSheet/BottomSheet';
@@ -20,7 +20,6 @@ import venuesInSolutionState from '../../atoms/venuesInSolutionState';
 import solutionState from '../../atoms/solutionState.js';
 import { useAppHistory } from '../../hooks/useAppHistory';
 import { useReset } from '../../hooks/useReset.js';
-import useMediaQuery from '../../hooks/useMediaQuery';
 import Sidebar from '../Sidebar/Sidebar';
 import useLocationForWayfinding from '../../hooks/useLocationForWayfinding';
 import locationIdState from '../../atoms/locationIdState';
@@ -100,6 +99,7 @@ MapTemplate.propTypes = {
     supportsUrlParameters: PropTypes.bool,
     center: PropTypes.string,
     useAppTitle: PropTypes.bool,
+    showMapMarkers: PropTypes.bool
 };
 
 /**
@@ -135,8 +135,9 @@ MapTemplate.propTypes = {
  * @param {boolean} [props.searchExternalLocations] - If you want to perform search for external locations in the Wayfinding mode. If set to true, Mapbox/Google places will be displayed depending on the Map Provider you are using. If set to false, the results returned will only be MapsIndoors results. The default is true.
  * @param {string} [props.center] - Specifies the coordinates where the map should load, represented as longitude and latitude values separated by a comma. If the specified coordinates intersect with a Venue, that Venue will be set as the current Venue.
  * @param {boolean} [props.useAppTitle] - Specifies if the Map Template should set the document title as defined in the App Config. The default value is set to false.
+ * @param {boolean} [props.showMapMarkers] - Specifies if the Map Template should show the base map providers Map Markers. The default value is set to true.
  */
-function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, primaryColor, logo, appUserRoles, directionsFrom, directionsTo, externalIDs, tileStyle, startZoomLevel, bearing, pitch, gmMapId, useMapProviderModule, kioskOriginLocationId, language, supportsUrlParameters, useKeyboard, timeout, miTransitionLevel, category, searchAllVenues, hideNonMatches, showRoadNames, showExternalIDs, searchExternalLocations, center, useAppTitle }) {
+function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, primaryColor, logo, appUserRoles, directionsFrom, directionsTo, externalIDs, tileStyle, startZoomLevel, bearing, pitch, gmMapId, useMapProviderModule, kioskOriginLocationId, language, supportsUrlParameters, useKeyboard, timeout, miTransitionLevel, category, searchAllVenues, hideNonMatches, showRoadNames, showExternalIDs, searchExternalLocations, center, useAppTitle, showMapMarkers }) {
 
     const [userSelectedLanguage, setUserSelectedLanguage] = useState(false);
     const [mapOptions, setMapOptions] = useState({ brandingColor: primaryColor });
@@ -200,7 +201,6 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
     const [, setPitch] = useRecoilState(pitchState);
 
     const isDesktop = useIsDesktop();
-    const isMobile = useMediaQuery('(max-width: 991px)');
     const resetState = useReset();
     const [pushAppView, goBack, currentAppView, currentAppViewPayload, appStates, resetAppHistory] = useAppHistory();
 
@@ -230,8 +230,8 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
 
             const miSdkApiTag = document.createElement('script');
             miSdkApiTag.setAttribute('type', 'text/javascript');
-            miSdkApiTag.setAttribute('src', 'https://app.mapsindoors.com/mapsindoors/js/sdk/4.41.1/mapsindoors-4.41.1.js.gz');
-            miSdkApiTag.setAttribute('integrity', 'sha384-MyVJoR6DbkW3ZM5QtEN7J6F9EpkBWSKjpMt4C0eK6UiHWHapyqQy1kPFrfTrjZaJ');
+            miSdkApiTag.setAttribute('src', 'https://app.mapsindoors.com/mapsindoors/js/sdk/4.41.2/mapsindoors-4.41.2.js.gz');
+            miSdkApiTag.setAttribute('integrity', 'sha384-glYwQ/XqeUiszPIc+4jYefSr6CaWnzhvsahILo2GvmumEuqAJm/aUtGn3ekm3GOX');
             miSdkApiTag.setAttribute('crossorigin', 'anonymous');
             document.body.appendChild(miSdkApiTag);
             miSdkApiTag.onload = () => {
@@ -498,12 +498,24 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
      */
     useEffect(() => {
         setMapOptions({
-            brandingColor: color, 
-            showRoadNames: showRoadNames, 
-            miTransitionLevel: miTransitionLevel, 
-            minZoom: ZoomLevelValues.minZoom
+            brandingColor: color,
+            // Ensure showRoadNames and showMapMarkers are booleans, even if appConfig.appSettings.showRoadNames/showMapMarkers is a string
+            showRoadNames:
+                showRoadNames !== undefined
+                    ? showRoadNames
+                    : (typeof appConfig?.appSettings?.showRoadNames === 'string'
+                        ? appConfig.appSettings.showRoadNames.toLowerCase() === 'true'
+                        : appConfig?.appSettings?.showRoadNames),
+            showMapMarkers:
+                showMapMarkers !== undefined
+                    ? showMapMarkers
+                    : (typeof appConfig?.appSettings?.showMapMarkers === 'string'
+                        ? appConfig.appSettings.showMapMarkers.toLowerCase() === 'true'
+                        : appConfig?.appSettings?.showMapMarkers),
+            miTransitionLevel: miTransitionLevel,
+            minZoom: ZoomLevelValues.minZoom,
         })
-    }, [primaryColor, showRoadNames, miTransitionLevel, color]);
+    }, [primaryColor, showRoadNames, miTransitionLevel, color, showMapMarkers, appConfig]);
 
     /*
      * React on changes in the start zoom level prop. If not defined, check if it is defined in app config.
@@ -782,7 +794,7 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
                         onRouteFinished={() => onRouteFinish()}
                     />
                 }
-                {isMobile &&
+                {!isDesktop &&
                     <BottomSheet
                         directionsFromLocation={directionsFromLocation}
                         directionsToLocation={directionsToLocation}