+++ packages/map-template/CHANGELOG.md
@@ -5,6 +5,73 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.82.0] - 2025-08-19
+
+### Added
+
+- Added lazy loading for Mapbox and Google Maps providers to improve initial page load performance
+
+## [1.81.6] - 2025-08-18
+
+### Fixed
+
+- Fixed extra media query in the `MapTemplate` component by using the useIsDesktop Hook instead
+
+## [1.81.5] - 2025-08-13
+
+### Changed
+
+- Removed Sentry from the map-template package
+
+## [1.81.4] - 2025-08-13
+
+### Fixed
+
+- Fixed a rare crash in the `OpeningHours` component when `startHours` was null by adding an additional null check
+
+## [1.81.3] - 2025-08-12
+
+### Fixed
+
+- Upgraded non-breaking depdenencies for the components package
+
+## [1.81.2] - 2025-08-05
+
+### Fixed
+
+- `showRoadNames` is configurable through the App Config object.
+- `showMapMarkers` is configurable through the App Config object.
+
+## [1.81.1] - 2025-08-05
+
+### Fixed
+
+- Fix removing `kioskOriginlocationId` display rule not reverting at runtime
+
+## [1.81.0] - 2025-08-04
+
+### Added
+
+- New `showMapMarkers` prop that specifies if the Google Maps/Mapbox specific map markers are shown or not.
+
+## [1.80.4] - 2025-07-31
+
+### Fixed
+
+- Fix gradients not showin in location details component
+
+## [1.80.3] - 2025-07-30
+
+### Fixed
+
+- Fix linting issues caused by the new lint rules
+
+## [1.80.2] - 2025-07-30
+
+### Fixed
+
+- Upgraded to SDK v4.41.2.
+
 ## [1.80.1] - 2025-07-03
 
 ### Fixed